backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
openai==1.50.0
anthropic==0.34.0
stripe==10.0.0
python-multipart==0.0.12
sse-starlette==2.1.0
redis==5.1.0
celery==5.4.0
sentry-sdk[fastapi]==2.14.0
posthog==3.6.0
python-dotenv==1.0.1
tenacity==9.0.0
tiktoken==0.7.0
aiofiles==24.1.0
Pillow==10.4.0
PyPDF2==3.0.1
python-docx==1.1.0
